Paper: Revert commit cd47c60
This is not a good change:
- First we explain how to establish unbiasedness (for the alg without Line 11).
- Then we generalize the proof to establish tail bounds.
- Then we show how to derive unbiasedness using the generalized result.
  - Note that at this stage unbiasedness is not even a motivated result (because we are not talking about an actually relevant algorithm.)

For readers who are reading the text the first time, they will get confused -- like, I thought we knew this already, why are we deriving it
agian. Or they will ask, why was the first proof presented.

The way it was presented in the previous version is intentional:

The direct proof for unbiasedness is easier to follow (without overwhelming the reader with notation) and then we can establish that the
functional invariant for the MGF works similarly.